[benchmark] Go over each hls benchmark and make sure important information are printed out to stdout
This is mainly to make sure a person can verify the application produces
the same output across record and replay.

+++ hdk/cl/examples/cl_hls/software/runtime/test_common.h
@@ -113,7 +113,7 @@ static inline void hls_peek_ocl64(uint64_t addr, uint64_t *data) {
     uint32_t lo, hi;
     hls_peek_ocl(addr, &lo);
     hls_peek_ocl(addr + 4, &hi);
-    *data = (hi << 32) | lo;
+    *data = ((uint64_t)(hi) << 32) | lo;
 }
 static inline void hls_poke_ocl64(uint64_t addr, uint64_t data) {
     hls_poke_ocl(addr, data & 0xffffffff);
